clinical-reconciliation/backend/requirements.txt
fastapi>=0.109.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.27.0
anthropic>=0.18.0
pydantic>=2.5.0
pyhealth>=1.1.5
pytest>=7.4.0
httpx>=0.25.0
